Chore: run-aws.sh 수정 (#88)
2024-04-15 13:07

run-aws.sh:
  - aws 서버 실행 스크립트 수정

ResumeControllerTest.java:
  - 테스트 코드 삭제

letterform.tsx & letter.d.ts:
  - 오류를 유발하는 자기소개서 기능 관련 코드 삭제

+++ data/run-aws.sh
@@ -1,4 +1,18 @@
 #!/bin/bash
+
+if ! dpkg -s build-essential; then
+  echo "[INFO] Package is not installed. Installation started."
+  apt update
+  apt install -y build-essential gfortran libblas-dev libopenblas-dev pkg-config
+  apt install -y default-libmysqlclient-dev libmariadb-dev
+  apt install -y default-jdk
+  python get-pip.py
+  pip install --no-cache-dir --upgrade -r ./requirements.txt
+  echo "[INFO] Package installation completed."
+else
+  echo "[INFO] Package is already installed."
+fi
+
+echo "[INFO] Starting FastAPI at port 8000."
 export PYTHONPATH=$PWD
-pip install --no-cache-dir --upgrade -r ./requirements.txt
 python -m uvicorn app.main:app --host "0.0.0.0" --port 8000
